[Krishna] | Added upload chunks endpoint
[Done]:
- Added file_exists attribute to chunk table which defaults to false
- Added upload chunks endpoint
- Modified u64 datatype to i64 for storing chunk size
- Stores actual chunk size in bigint downcasting u64 to i64

[Next Steps]:
- Can check for SHA after file is added to notify client for integrity check

+++ src/db/models.rs
@@ -36,7 +36,7 @@ pub struct Chunk {
     pub id: i32,
     pub index_id: i32,
     pub name: String,
-    pub size: i32,
+    pub size: i64,
     pub creation_time: DateTime<Utc>,
     pub accessed_time: DateTime<Utc>,
     pub tags: Option<Vec<i32>>,
